third_party/WebKit/LayoutTests/accessibility/textarea-selection-expected.txt

This tests that text selection is reported correctly for textarea elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS textareaAccessible.selectionStart is 0
PASS textareaAccessible.selectionEnd is 25
PASS textareaAccessible.selectionStartLineNumber is 0
PASS textareaAccessible.selectionEndLineNumber is 3
PASS textareaAccessible.selectionStart is 0
PASS textareaAccessible.selectionEnd is 7
PASS textareaAccessible.selectionStartLineNumber is 0
PASS textareaAccessible.selectionEndLineNumber is 1
PASS textareaAccessible.selectionStart is 0
PASS textareaAccessible.selectionEnd is 7
PASS textareaAccessible.selectionStartLineNumber is 0
PASS textareaAccessible.selectionEndLineNumber is 1
PASS textareaAccessible.selectionStart is 7
PASS textareaAccessible.selectionEnd is 14
PASS textareaAccessible.selectionStartLineNumber is 1
PASS textareaAccessible.selectionEndLineNumber is 2
PASS textareaAccessible.selectionStart is 7
PASS textareaAccessible.selectionEnd is 14
PASS textareaAccessible.selectionStartLineNumber is 1
PASS textareaAccessible.selectionEndLineNumber is 2
PASS textareaAccessible.selectionStart is 14
PASS textareaAccessible.selectionEnd is 21
PASS textareaAccessible.selectionStartLineNumber is 2
PASS textareaAccessible.selectionEndLineNumber is 3
PASS textareaAccessible.selectionStart is 14
PASS textareaAccessible.selectionEnd is 21
PASS textareaAccessible.selectionStartLineNumber is 2
PASS textareaAccessible.selectionEndLineNumber is 3
PASS textareaAccessible.selectionStart is 14
PASS textareaAccessible.selectionEnd is 21
PASS textareaAccessible.selectionStartLineNumber is 2
PASS textareaAccessible.selectionEndLineNumber is 3
PASS emptyTextareaAccessible.selectionStart is 0
PASS emptyTextareaAccessible.selectionEnd is 0
PASS emptyTextareaAccessible.selectionStartLineNumber is 0
PASS emptyTextareaAccessible.selectionEndLineNumber is 0
PASS emptyTextareaAccessible.selectionStart is 0
PASS emptyTextareaAccessible.selectionEnd is 0
PASS emptyTextareaAccessible.selectionStartLineNumber is 0
PASS emptyTextareaAccessible.selectionEndLineNumber is 0
PASS successfullyParsed is true

TEST COMPLETE

 
